refactor: likeProduct
