+++ services/validators/like-resource-validator.js
@@ -8,8 +8,6 @@ const { Product, Like } = require('../../db/models')
 
 class LikeResourceValidator {
   static async likeProduct(req) {
-    // current login user
-    const loginUser = AuthToolKit.getUser(req)
     const { productId } = req.params
 
     // check whether a specific product exists
@@ -19,17 +17,8 @@ class LikeResourceValidator {
     if (!product) {
       throw new APIError({ code: code.NOTFOUND, message: '找不到對應項目' })
     }
-    // check whether the user has repeatedly liked the same product
-    // return error if it's true
-    const findLikeOption = {
-      where: { userId: loginUser.id, productId }
-    }
 
-    const [like, created] = await Like.findOrCreate(findLikeOption)
-    if (!created) {
-      throw new APIError({ code: code.FORBIDDEN, message: '使用者不能重複喜歡同個產品' })
-    }
-    const resultData = like
+    const resultData = null
     return { data: resultData }
   }
